Remove optional telemetry integration
