+++ THIRD_PARTY_NOTICES.md
@@ -7,7 +7,6 @@ records the exact resolved dependency set used for this release.
 | Dependency group | Packages and licenses |
 | --- | --- |
 | Base | Click (BSD-3-Clause), Jinja2 (BSD-3-Clause), Pydantic (MIT), PyYAML (MIT), Rich (MIT) |
-| LLM | Anthropic (MIT), Boto3 (Apache-2.0), LiteLLM (MIT), OpenAI (Apache-2.0), Protobuf (BSD-3-Clause) |
+| LLM | Anthropic (MIT), Boto3 (Apache-2.0), LiteLLM (MIT), OpenAI (Apache-2.0) |
 | Tier 3 | Harbor (Apache-2.0) |
-| Telemetry | OpenTelemetry API, SDK, and OTLP HTTP Exporter (Apache-2.0) |
 | Security | Bandit (Apache-2.0), pip-audit (Apache-2.0) |
