+++ docs/configuration.mdx
@@ -1,6 +1,6 @@
 ---
 title: "Configuration"
-description: "The credential map plus provider, embeddings, and telemetry configuration for SkillEvaluator."
+description: "The credential map plus provider and embeddings configuration for SkillEvaluator."
 ---
 
 ## What needs credentials, and what runs offline
@@ -166,15 +166,5 @@ host-process configuration. It cannot reference protected host credentials
 under an innocent alias.
 Names that control the launcher or dynamic runtime, including `PATH`,
 `PYTHON*`, `LD_*`, `DYLD_*`, `DOCKER_*`, `COMPOSE_*`, proxy variables, and
-telemetry exporters, are rejected. Configure the selected Harbor backend in
+tracing exporters, are rejected. Configure the selected Harbor backend in
 the host environment instead.
-
-## Telemetry
-
-Telemetry is disabled by default. To opt in, install
-`skillevaluator[telemetry]`, set `SKILLEVALUATOR_TELEMETRY_ENABLED=true`, and
-configure a standard `OTEL_EXPORTER_OTLP_ENDPOINT`. Identity export is
-controlled by `SKILLEVALUATOR_TELEMETRY_IDENTITY_MODE`: the default
-`team_only` exports no user identity, `hashed` exports only a hash, and any
-other value exports the raw login — leave it unset unless you need
-attribution.
