+++ docs/tier3-live-evaluation.mdx
@@ -189,7 +189,7 @@ Security notes:
   provider routing, backend controls, and aliases such as
   `SAFE_NAME=${NVIDIA_API_KEY}` are rejected.
 - `runtime_env` cannot override host launcher, language-runtime, Docker,
-  Compose, proxy, or telemetry controls. Configure the selected backend in the
+  Compose, proxy, or tracing controls. Configure the selected backend in the
   host environment.
 
 ## Commands
@@ -581,7 +581,7 @@ required by the selected agent and environment.
 
 ## See Also
 
-- [Configuration](configuration.mdx) — provider, embedding, credential, and telemetry setup.
+- [Configuration](configuration.mdx) — provider, embedding, and credential setup.
 - [Installation](installation.mdx) — extras and system requirements.
 - [Tier 1 validation](tier1-validation.mdx) — deterministic and security checks.
 - [Tier 2 deduplication](tier2-deduplication.mdx) — semantic overlap checks.
