+++ pyproject.toml
@@ -34,8 +34,8 @@ license-files = [
     "NOTICE",
     "THIRD_PARTY_NOTICES.md",
 ]
-# Base install is static-validation capable (Tier 1) and must NOT pull Harbor,
-# LLM, or OpenTelemetry. Those live behind the extras below.
+# Base install is static-validation capable (Tier 1) and must NOT pull Harbor
+# or LLM clients. Those live behind the extras below.
 dependencies = [
     "click>=8.3.3,<9",
     "jinja2>=3.1",
@@ -51,8 +51,6 @@ llm = [
     "boto3>=1.34.0",
     "litellm>=1.83.10,<1.89.0.dev0",
     "openai>=2.21.0",
-    # CVE-2026-0994 / BDSA-2025-123809 complete recursion-limit fix.
-    "protobuf>=7.35.1",
 ]
 # Tier 2 intra-skill deduplication and local-catalog inter-skill similarity.
 tier2 = [
@@ -67,14 +65,6 @@ tier3 = [
     # CVE-2026-48522 through CVE-2026-48526 fixed floor.
     "pyjwt[crypto]>=2.13.0",
 ]
-# OpenTelemetry export (optional; telemetry degrades gracefully when absent).
-telemetry = [
-    "opentelemetry-api>=1.37.0",
-    "opentelemetry-exporter-otlp-proto-http>=1.37.0",
-    "opentelemetry-sdk>=1.37.0",
-    # CVE-2026-0994 / BDSA-2025-123809 complete recursion-limit fix.
-    "protobuf>=7.35.1",
-]
 # Static security scanners (floors raised to the established release baseline).
 security = [
     "bandit>=1.9.4",
@@ -82,7 +72,7 @@ security = [
 ]
 # Full public install convenience bundle.
 all = [
-    "skillevaluator[tier2,tier3,telemetry,security]",
+    "skillevaluator[tier2,tier3,security]",
 ]
 dev = [
     # PyPA build backs the `make build` target.
