+++ tests/test_harbor_runner_environment.py
@@ -424,8 +424,6 @@ def test_run_harbor_eval_stages_per_agent_credential_trees(
         lambda **_kwargs: {"execution_status": "complete", "execution_errors": [], "metrics": [], "agents": {}},
     )
     monkeypatch.setattr(runner, "render_agent_eval_html_report", lambda *_args, **_kwargs: tmp_path / "report.html")
-    monkeypatch.setattr(runner, "record_agent_eval_summary", lambda **_kwargs: None)
-
     result = runner.run_harbor_eval(
         skill,
         ["opencode", "claude-code"],
@@ -488,7 +486,6 @@ def test_harbor_subprocess_environment_excludes_arbitrary_host_secrets(
         "TMPDIR": str(tmp_path / "tmp"),
         "LANG": "C.UTF-8",
         "DOCKER_HOST": "unix:///safe/docker.sock",
-        "SKILLEVALUATOR_TELEMETRY_DISABLED": "true",
         "SERVICE_TOKEN": "declared-runtime-secret",
         **provider_env,
     }
